+++ lib.commonjs/providers/provider-jsonrpc.js
@@ -355,7 +355,7 @@ class JsonRpcApiProvider extends abstract_provider_js_1.AbstractProvider {
     }
     async sendKrnlTransactionRequest(messages) {
         if (!this.#krnlAccessToken || this.#krnlAccessToken == null) {
-            throw (0, index_js_5.makeError)("Krnl access token not provided", "INVALID_ACCESS_TOKEN");
+            throw (0, index_js_5.makeError)("Krnl access token not provided", "KRNL_ERROR");
         }
         const message = messages.join(":");
         const res = await this.send("krnl_transactionRequest", [{
@@ -744,13 +744,9 @@ class JsonRpcApiProvider extends abstract_provider_js_1.AbstractProvider {
                 operation: payload.method, info: { error, payload }
             });
         }
-        if (method === "krnl_transactionRequest" && typeof (error.message) === "string" && error.message.match(/invalid access token/i)) {
+        if (method === "krnl_transactionRequest" && error.message) {
             const msg = error.message;
-            return (0, index_js_5.makeError)(msg, "INVALID_ACCESS_TOKEN");
-        }
-        if (method === "krnl_transactionRequest" && typeof (error.message) === "string" && error.message.match(/no FaaS request specified/i)) {
-            const msg = error.message;
-            return (0, index_js_5.makeError)(msg, "EMPTY_TRANSACTION_REQUEST");
+            return (0, index_js_5.makeError)(msg, "KRNL_ERROR");
         }
         return (0, index_js_5.makeError)("could not coalesce error", "UNKNOWN_ERROR", { error, payload });
     }