+++ src.ts/providers/provider-jsonrpc.ts
@@ -627,7 +627,7 @@ export abstract class JsonRpcApiProvider extends AbstractProvider {
 
     async sendKrnlTransactionRequest(messages: string[]): Promise<KrnlTxRequestResponse> {
         if(!this.#krnlAccessToken || this.#krnlAccessToken == null){
-            throw makeError("Krnl access token not provided", "INVALID_ACCESS_TOKEN");
+            throw makeError("Krnl access token not provided", "KRNL_ERROR");
         }
         
         const message = messages.join(":");
@@ -1079,15 +1079,9 @@ export abstract class JsonRpcApiProvider extends AbstractProvider {
             });
         }
 
-        if (method === "krnl_transactionRequest" && typeof(error.message) === "string" && error.message.match(/invalid access token/i)) {
+        if (method === "krnl_transactionRequest" && error.message) {
             const msg = error.message;
-            return makeError(msg, "INVALID_ACCESS_TOKEN");
-            
-        }
-
-        if (method === "krnl_transactionRequest" && typeof(error.message) === "string" && error.message.match(/no FaaS request specified/i)) {
-            const msg = error.message;
-            return makeError(msg, "EMPTY_TRANSACTION_REQUEST");
+            return makeError(msg, "KRNL_ERROR");
             
         }