+++ CHANGELOG.md
@@ -3,6 +3,10 @@ Change Log
 
 This change log is maintained by `src.ts/_admin/update-changelog.ts` but may also be manually updated.
 
+krnl/v1.0.2(2024-02-14 15:07)
+---------------------------------
+  - generalised krnl errors
+
 krnl/v1.0.1 (2024-02-14 14:29)
 ---------------------------------
   - add support for empty krnl transaction request